+++ packages/dotcom/tsconfig.types.json
@@ -10,5 +10,6 @@
         "allowJs": false,
         "emitDeclarationOnly": true
     },
-    "exclude": ["**/*.test.ts", "dist/**/*"]
+    "exclude": ["**/*.test.ts", "dist/**/*"],
+    "files": ["./src/index.ts"]
 }
