+++ packages/shared/src/lib/tracking.ts
@@ -1,8 +1,8 @@
-import { EpicTest, EpicVariant } from '../types/epic';
 import { BannerTest, BannerVariant } from '../types/banner';
 import { OphanAction, OphanComponentEvent } from '../types/ophan';
 import { addRegionIdToSupportUrl } from './geolocation';
 import { Tracking } from '../types/shared';
+import { EpicTest, EpicVariant } from '../types/epic/test';
 
 type LinkParams = {
     REFPVID: string;
