--- packages/shared/src/types/epic.ts
@@ -1,239 +1,0 @@
-import { OphanComponentEvent } from './ophan';
-import {
-    ArticlesViewedSettings,
-    ArticleCounts,
-    UserCohort,
-    Cta,
-    SecondaryCta,
-    Stage,
-    Test,
-    TickerSettings,
-    Variant,
-    WeeklyArticleHistory,
-    tickerSettingsSchema,
-    ctaSchema,
-    Tracking,
-    PageTracking,
-    trackingSchema,
-    secondaryCtaSchema,
-    ArticlesViewedByTagSettings,
-} from './shared';
-import { ReminderFields } from '../lib/reminderFields';
-import { CountryGroupId } from '../lib/geolocation';
-import { z } from 'zod';
-
-export type Tag = {
-    id: string;
-    type: string;
-};
-
-interface View {
-    date: number;
-    testId: string;
-}
-export type ViewLog = View[];
-
-export type EpicTargeting = {
-    contentType: string;
-    sectionName?: string; // Deprecated - use sectionId
-    sectionId?: string;
-    shouldHideReaderRevenue: boolean;
-
-    // TODO let's replace these with Design Type/a single property after migration
-    isMinuteArticle: boolean;
-    isPaidContent: boolean;
-
-    tags: Tag[];
-    mvtId?: number;
-    epicViewLog?: ViewLog;
-    countryCode?: string;
-    weeklyArticleHistory?: WeeklyArticleHistory;
-    hasOptedOutOfArticleCount: boolean;
-    showSupportMessaging: boolean;
-    isRecurringContributor: boolean;
-    lastOneOffContributionDate?: number; // Platform to send undefined or a timestamp date
-    modulesVersion?: string;
-    url?: string;
-    browserId?: string; // Only present if the user has consented to browserId-based targeting
-};
-
-export type EpicPayload = {
-    tracking: PageTracking;
-    targeting: EpicTargeting;
-};
-
-export type EpicType = 'ARTICLE' | 'LIVEBLOG';
-
-const articleCountsSchema = z.object({
-    for52Weeks: z.number(),
-    forTargetedWeeks: z.number(),
-});
-
-export interface MaxViews {
-    maxViewsCount: number;
-    maxViewsDays: number;
-    minDaysBetweenViews: number;
-}
-
-const maxViewsSchema = z.object({
-    maxViewsCount: z.number(),
-    maxViewsDays: z.number(),
-    minDaysBetweenViews: z.number(),
-});
-
-export interface SeparateArticleCount {
-    type: 'above';
-}
-
-const separateArticleCountSchema = z.object({
-    type: z.string(),
-});
-
-const reminderFieldsSchema = z.object({
-    reminderCta: z.string(),
-    reminderLabel: z.string(),
-    reminderPeriod: z.string(),
-});
-
-export interface EpicVariant extends Variant {
-    name: string;
-    heading?: string;
-    paragraphs: string[];
-    highlightedText?: string;
-    tickerSettings?: TickerSettings;
-    cta?: Cta;
-    secondaryCta?: SecondaryCta;
-    footer?: string;
-    backgroundImageUrl?: string;
-    showReminderFields?: ReminderFields;
-    modulePathBuilder?: (version?: string) => string;
-    separateArticleCount?: SeparateArticleCount;
-    showChoiceCards?: boolean;
-    choiceCardAmounts?: ChoiceCardAmounts;
-    defaultChoiceCardFrequency?: ContributionFrequency;
-
-    // Variant level maxViews are for special targeting tests. These
-    // are handled differently to our usual copy/design tests. To
-    // set up a targeting test, the test should be set to alwaysAsk
-    // and each variant should have a maxViews set. We then check if a
-    // a user should actually see an epic after they have been assigned to
-    // the test + variant. This means users **wont** fall through to a test
-    // with lower priority.
-    maxViews?: MaxViews;
-    showSignInLink?: boolean;
-}
-
-const variantSchema = z.object({
-    name: z.string(),
-    heading: z.string().optional(),
-    paragraphs: z.array(z.string()),
-    highlightedText: z.string().optional(),
-    tickerSettings: tickerSettingsSchema.optional(),
-    cta: ctaSchema.optional(),
-    secondaryCta: secondaryCtaSchema.optional(),
-    footer: z.string().optional(),
-    backgroundImageUrl: z.string().optional(),
-    showReminderFields: reminderFieldsSchema.optional(),
-    separateArticleCount: separateArticleCountSchema.optional(),
-    maxViews: maxViewsSchema.optional(),
-    showSignInLink: z.boolean().optional(),
-});
-
-export type EpicProps = {
-    variant: EpicVariant;
-    tracking: Tracking;
-    countryCode?: string;
-    articleCounts: ArticleCounts;
-    // eslint-disable-next-line @typescript-eslint/ban-types
-    onReminderOpen?: Function;
-    email?: string;
-    fetchEmail?: () => Promise<string | null>;
-    submitComponentEvent?: (componentEvent: OphanComponentEvent) => void;
-    openCmp?: () => void;
-    hasConsentForArticleCount?: boolean;
-    stage?: Stage;
-};
-
-export const epicPropsSchema = z.object({
-    variant: variantSchema,
-    tracking: trackingSchema,
-    countryCode: z.string().optional(),
-    articleCounts: articleCountsSchema,
-    onReminderOpen: z.any().optional(),
-    email: z.string().optional(),
-    fetchEmail: z.any().optional(),
-    submitComponentEvent: z.any().optional(),
-    openCmp: z.any().optional(),
-    hasConsentForArticleCount: z.boolean().optional(),
-    stage: z.string().optional(),
-});
-
-interface ControlProportionSettings {
-    proportion: number;
-    offset: number;
-}
-
-export type ContributionFrequency = 'ONE_OFF' | 'MONTHLY' | 'ANNUAL';
-
-export interface AmountSelection {
-    amounts: number[];
-    defaultAmount: number;
-}
-
-export type ContributionAmounts = {
-    [key in ContributionFrequency]: AmountSelection;
-};
-
-export interface AmountsTestVariant {
-    name: string;
-    amounts: ContributionAmounts;
-}
-
-export interface AmountsTest {
-    name: string;
-    isLive: boolean;
-    variants: AmountsTestVariant[];
-    seed: number;
-}
-
-export type ConfiguredRegionAmounts = {
-    control: ContributionAmounts;
-    test?: AmountsTest;
-};
-
-export type ChoiceCardAmounts = {
-    [key in CountryGroupId]: ConfiguredRegionAmounts;
-};
-
-export interface EpicTest extends Test<EpicVariant> {
-    name: string;
-    isOn: boolean;
-    locations: CountryGroupId[];
-    tagIds: string[];
-    sections: string[]; // section IDs
-    excludedTagIds: string[];
-    excludedSections: string[];
-    alwaysAsk: boolean;
-    maxViews?: MaxViews;
-    userCohort: UserCohort;
-    isLiveBlog: boolean;
-    hasCountryName: boolean;
-    variants: EpicVariant[];
-    highPriority: boolean;
-    useLocalViewLog: boolean;
-    articlesViewedSettings?: ArticlesViewedSettings;
-    articlesViewedByTagSettings?: ArticlesViewedByTagSettings;
-    hasArticleCountInCopy: boolean;
-
-    audience?: number;
-    audienceOffset?: number;
-
-    // These are specific to hardcoded tests
-    expiry?: string;
-    campaignId?: string;
-
-    controlProportionSettings?: ControlProportionSettings;
-
-    isSuperMode?: boolean;
-    canShow?: (targeting: EpicTargeting) => boolean;
-}
