+++ packages/shared/src/types/epic/index.ts
@@ -1,0 +1,3 @@
+export * from './payload';
+export * from './test';
+export * from './props';
