+++ packages/shared/src/types/shared/index.ts
@@ -1,0 +1,3 @@
+export * from './props';
+export * from './payload';
+export * from './test';
