+++ packages/shared/src/types/shared/payload.ts
@@ -1,0 +1,18 @@
+export type PageTracking = {
+    ophanPageId: string;
+    platformId: string;
+    referrerUrl: string;
+    clientName: string;
+};
+
+export type TagCounts = {
+    [tag: string]: number;
+};
+
+export type WeeklyArticleLog = {
+    week: number;
+    count: number;
+    tags?: TagCounts;
+};
+
+export type WeeklyArticleHistory = WeeklyArticleLog[];
